% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_time_series.R
\name{acf_ts}
\alias{acf_ts}
\title{Fonction d'autocorrelation (ACF)}
\usage{
acf_ts(x, lag.max = 10L)
}
\arguments{
\item{x}{serie ; @param lag.max retard maximal.}
}
\value{
vecteur des autocorrelations (retards 0 a \code{lag.max}).
}
\description{
\eqn{\hat\rho_k = \hat\gamma_k/\hat\gamma_0}, \eqn{\hat\gamma_k=\frac1n\sum_{t}
(x_t-\bar x)(x_{t-k}-\bar x)}.
}
