% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_did.R
\name{aggregate_att}
\alias{aggregate_att}
\title{Agrégation des ATT(g,t) (Callaway-Sant'Anna)}
\usage{
aggregate_att(attgt, type = c("simple", "dynamic", "group"))
}
\arguments{
\item{attgt}{sortie de \code{att_gt}.}

\item{type}{"simple" (ATT global, périodes post, pondéré par taille de groupe),
"dynamic" (event-study par ancienneté \eqn{e=t-g}) ou "group" (par cohorte).}
}
\value{
pour "simple" un scalaire ; sinon un data.frame (\code{e}/\code{g}, \code{att}).
}
\description{
Agrégation des ATT(g,t) (Callaway-Sant'Anna)
}
