% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_time_series.R
\name{ar_yw}
\alias{ar_yw}
\title{Estimation d'un AR(p) par les equations de Yule-Walker}
\usage{
ar_yw(x, order = 1L)
}
\arguments{
\item{x}{serie}

\item{order}{ordre p}
}
\value{
liste : \code{ar} (coefficients), \code{var_pred} (variance d'innovation), \code{mean}.
}
\description{
Resout \eqn{R\phi=r} (R matrice de Toeplitz des autocorrelations). Estimateur
de moments, coherent et rapide.
}
