% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/45_garch.R
\name{arch_lm_test}
\alias{arch_lm_test}
\title{Test ARCH-LM (heteroscedasticite conditionnelle)}
\usage{
arch_lm_test(x, q = 5L)
}
\arguments{
\item{x}{serie (rendements)}

\item{q}{nombre de retards.}
}
\value{
liste : \code{statistic}, \code{df}, \code{p_value}.
}
\description{
Regresse \eqn{\hat\varepsilon_t^2} sur ses \eqn{q} retards ; \eqn{nR^2\sim
\chi^2_q} sous absence d'effet ARCH. Rejet = la volatilite se regroupe.
}
