% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_time_series.R
\name{arma_css}
\alias{arma_css}
\title{Estimation d'un ARMA(p,q) par moindres carres conditionnels (CSS)}
\usage{
arma_css(x, p = 1L, q = 1L)
}
\arguments{
\item{x}{serie}

\item{p, q}{ordres AR et MA}
}
\value{
liste : \code{ar}, \code{ma}, \code{mean}, \code{sigma2}.
}
\description{
Minimise \eqn{\sum_t \hat\varepsilon_t^2}, ou \eqn{\hat\varepsilon_t=(x_t-\mu)-
\sum \phi_i(x_{t-i}-\mu)-\sum\theta_j\hat\varepsilon_{t-j}} est calcule
recursivement (erreurs initiales nulles). Optimisation via \code{optim}.
}
