% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_linalg.R
\name{back_substitution}
\alias{back_substitution}
\title{Résolution d'un système triangulaire supérieur par remontée}
\usage{
back_substitution(U, b)
}
\arguments{
\item{U}{matrice triangulaire supérieure (n x n), diagonale non nulle.}

\item{b}{vecteur second membre (longueur n).}
}
\value{
le vecteur solution x.
}
\description{
Résout \eqn{U x = b} avec \eqn{U} triangulaire supérieure. Brique de la
résolution MCO par QR (remontée sur \eqn{R_1}, éq. 0.8) et par Cholesky
(éq. 0.10).
}
