% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35_nn_training.R
\name{batch_norm_backward}
\alias{batch_norm_backward}
\title{Batch normalization (passe arriere)}
\usage{
batch_norm_backward(dout, cache)
}
\arguments{
\item{dout}{gradient en sortie (n x d).}

\item{cache}{sortie de \code{batch_norm}.}
}
\value{
liste : \code{dX}, \code{dgamma}, \code{dbeta}.
}
\description{
Gradients de la perte par rapport a \eqn{X}, \eqn{\gamma}, \eqn{\beta} etant
donne \code{dout} (gradient en sortie).
}
