% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_diagnostics.R
\name{bg_test}
\alias{bg_test}
\title{Test de Breusch-Godfrey (autocorrelation d'ordre p, LM)}
\usage{
bg_test(formula, data, order = 1)
}
\arguments{
\item{formula, data}{cf. \code{bp_test}}

\item{order}{ordre du retard}
}
\value{
liste : \code{statistic}, \code{df}, \code{p_value}.
}
\description{
Regression des residus sur les regresseurs ET \eqn{p} retards des residus ;
\eqn{n R^2 \sim \chi^2_p}. Plus general que Durbin-Watson (ordre eleve,
regresseurs retardes admis).
}
