% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_boosting.R
\name{boost_loss_path}
\alias{boost_loss_path}
\title{Trajectoire de la perte d'entraînement/test selon le nombre d'arbres}
\usage{
boost_loss_path(object, data, y)
}
\arguments{
\item{object}{objet \code{boost}.}

\item{data}{data.frame des prédicteurs.}

\item{y}{réponse correspondante.}
}
\value{
vecteur de perte (moyenne de \eqn{\ell}) après m arbres, m = 1..M.
}
\description{
Trajectoire de la perte d'entraînement/test selon le nombre d'arbres
}
