% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_diagnostics.R
\name{bp_test}
\alias{bp_test}
\title{Test de Breusch-Pagan (heteroscedasticite)}
\usage{
bp_test(formula, data)
}
\arguments{
\item{formula}{formule du modele}

\item{data}{data.frame}
}
\value{
liste : \code{statistic}, \code{df}, \code{p_value}.
}
\description{
Version studentisee de Koenker : \eqn{n R^2} de la regression des residus au
carre sur les regresseurs. \eqn{H_0} : homoscedasticite. Statistique
\eqn{\sim \chi^2_{p-1}}.
}
