% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_nonparametric.R
\name{bw_loocv}
\alias{bw_loocv}
\title{Selection de fenetre par validation croisee leave-one-out (Nadaraya-Watson)}
\usage{
bw_loocv(x, y, bws)
}
\arguments{
\item{x, y}{donnees}

\item{bws}{fenetres candidates}
}
\value{
liste : \code{bw} (optimale), \code{cv} (erreurs CV par fenetre).
}
\description{
Minimise \eqn{\sum_i (y_i-\hat m_{-i}(x_i))^2}, ou \eqn{\hat m_{-i}} exclut
l'observation \eqn{i} (formule du residu : retirer le poids diagonal).
}
