% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/36_cnn.R
\name{conv2d}
\alias{conv2d}
\title{Convolution 2D (correlation croisee, mode "valid")}
\usage{
conv2d(X, K, b = rep(0, dim(K)[3]))
}
\arguments{
\item{X}{image d'entree (matrice H x W).}

\item{K}{noyaux, tableau (kh x kw x F)}

\item{b}{biais (longueur F).}
}
\value{
liste : \code{out} (Hout x Wout x F), \code{cache}.
}
\description{
Pour chaque filtre \eqn{f}, \eqn{Y_{ij}^f=b_f+\sum_{a,b}X_{i+a-1,j+b-1}K_{ab}^f}.
Partage de poids : le meme noyau \eqn{K^f} est applique en toute position.
}
