% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/36_cnn.R
\name{conv2d_backward}
\alias{conv2d_backward}
\title{Retropropagation de la convolution 2D}
\usage{
conv2d_backward(dout, cache)
}
\arguments{
\item{dout}{gradient en sortie (Hout x Wout x F).}

\item{cache}{sortie de \code{conv2d}.}
}
\value{
liste : \code{dX}, \code{dK}, \code{db}.
}
\description{
Gradients par rapport a l'entree, aux noyaux et aux biais, etant donne \code{dout}.
}
