% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_cart.R
\name{cost_complexity_prune}
\alias{cost_complexity_prune}
\title{Élagage coût-complexité (éq. 8.4-8.5)}
\usage{
cost_complexity_prune(object, alpha)
}
\arguments{
\item{object}{objet \code{cart}.}

\item{alpha}{coût par feuille \eqn{\alpha \ge 0}.}
}
\value{
objet \code{cart} élagué.
}
\description{
Effondre récursivement les sous-arbres dont le maintien n'améliore pas
\eqn{R_\alpha(T)=R(T)+\alpha|T|} (weakest-link, Prop. 8.2).
}
