% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41_clustering.R
\name{dbscan_fit}
\alias{dbscan_fit}
\title{DBSCAN (clustering base sur la densite)}
\usage{
dbscan_fit(X, eps, minPts = 5L)
}
\arguments{
\item{X}{matrice n x p ; @param eps rayon ; @param minPts densite minimale.}
}
\value{
vecteur d'etiquettes (0 = bruit).
}
\description{
Un point est \strong{coeur} s'il a au moins \code{minPts} voisins dans un rayon \code{eps}.
Les clusters sont des composantes connexes de points-coeur (et leurs voisins) ;
les points isoles sont du \strong{bruit} (label 0). Detecte des formes arbitraires,
sans fixer le nombre de clusters.
}
