% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_knn.R
\name{edge_length}
\alias{edge_length}
\title{Longueur d'arête d'un voisinage capturant une fraction r (éq. 7.3)}
\usage{
edge_length(r, p)
}
\arguments{
\item{r}{fraction des données à capturer (dans (0,1]).}

\item{p}{dimension.}
}
\value{
la longueur d'arête nécessaire.
}
\description{
\eqn{e_p(r) = r^{1/p}} dans l'hypercube unité.
}
