% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_mcmc.R
\name{ess}
\alias{ess}
\title{Taille d'echantillon effective (ESS)}
\usage{
ess(x)
}
\arguments{
\item{x}{chaine (vecteur).}
}
\value{
la taille d'echantillon effective.
}
\description{
\eqn{\mathrm{ESS}=n/(1+2\sum_k\hat\rho_k)}, somme des autocorrelations tronquee
au premier terme negatif (sequence positive initiale de Geyer). Mesure combien
de tirages INDEPENDANTS equivalent a la chaine correlee.
}
