% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_diagnostics.R
\name{fgls}
\alias{fgls}
\title{Moindres carres generalises FAISABLES (FGLS) pour heteroscedasticite}
\usage{
fgls(formula, data)
}
\arguments{
\item{formula, data}{cf. \code{bp_test}.}
}
\value{
liste : \code{coefficients}, \code{weights}, \code{se} (classiques WLS).
}
\description{
Modele de variance : \eqn{\log\hat\varepsilon^2 = X\gamma}. On en deduit des
poids \eqn{\hat w_i = 1/\exp(X_i\hat\gamma)} et l'on refait une WLS (Module 2).
Plus efficace que l'OLS sous heteroscedasticite bien specifiee.
}
