% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_linalg.R
\name{forward_substitution}
\alias{forward_substitution}
\title{Résolution d'un système triangulaire inférieur par descente}
\usage{
forward_substitution(L, b)
}
\arguments{
\item{L}{matrice triangulaire inférieure (n x n), diagonale non nulle.}

\item{b}{vecteur second membre (longueur n).}
}
\value{
le vecteur solution x.
}
\description{
Résout \eqn{L x = b} avec \eqn{L} triangulaire inférieure. Brique de la
résolution des équations normales par Cholesky, éq. (0.10) de
derivations/00_linalg.qmd.
}
