% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/45_garch.R
\name{garch_fit}
\alias{garch_fit}
\title{Estimation d'un GARCH(1,1) par (quasi-)maximum de vraisemblance}
\usage{
garch_fit(x, maxit = 500L)
}
\arguments{
\item{x}{serie (rendements centres)}

\item{maxit}{iterations de l'optimiseur.}
}
\value{
liste : \code{omega}, \code{alpha}, \code{beta}, \code{sigma} (volatilite conditionnelle), \code{loglik}.
}
\description{
\eqn{x_t=\sigma_t z_t}, \eqn{z_t\sim(0,1)}, avec la variance conditionnelle
\eqn{\sigma_t^2=\omega+\alpha x_{t-1}^2+\beta\sigma_{t-1}^2}. On maximise la
log-vraisemblance gaussienne (paramétrage assurant \eqn{\omega>0}, \eqn{\alpha,
\beta\ge0}, \eqn{\alpha+\beta<1} : stationnarite).
}
