% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_validation.R
\name{gaussian_loglik}
\alias{gaussian_loglik}
\title{Log-vraisemblance gaussienne (régression linéaire) au MLE}
\usage{
gaussian_loglik(rss, n)
}
\arguments{
\item{rss}{somme des carrés des résidus.}

\item{n}{taille d'échantillon.}
}
\value{
la log-vraisemblance maximisée.
}
\description{
Log-vraisemblance gaussienne (régression linéaire) au MLE
}
