% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_validation.R
\name{gcv_linear}
\alias{gcv_linear}
\title{Validation croisée généralisée (GCV)}
\usage{
gcv_linear(X, y)
}
\arguments{
\item{X}{matrice de design (constante incluse).}

\item{y}{réponse.}
}
\value{
liste : \code{gcv}.
}
\description{
Remplace \eqn{h_{ii}} par la moyenne \eqn{\bar h = p/n} dans la formule LOOCV.
}
