% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_mcmc.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{Diagnostic de convergence de Gelman-Rubin (statistique R-hat)}
\usage{
gelman_rubin(chains)
}
\arguments{
\item{chains}{liste de chaines (vecteurs de meme longueur).}
}
\value{
la statistique R-hat.
}
\description{
Compare la variance INTER-chaines a la variance INTRA-chaine :
\eqn{\hat R=\sqrt{\hat V/W}}, \eqn{\hat V=\frac{n-1}n W+\frac Bn}. \eqn{\hat R\to
1} a convergence ; \eqn{>1.1} signale une non-convergence.
}
