% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_pca_kmeans_em.R
\name{gmm_loglik}
\alias{gmm_loglik}
\title{Log-vraisemblance observée d'un mélange gaussien}
\usage{
gmm_loglik(X, pi, mu, Sigma)
}
\arguments{
\item{X}{données n x p.}

\item{pi}{poids (longueur K).}

\item{mu}{liste (ou matrice K x p) des moyennes.}

\item{Sigma}{liste des K covariances.}
}
\value{
la log-vraisemblance observée.
}
\description{
Log-vraisemblance observée d'un mélange gaussien
}
