% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/44_var.R
\name{granger_test}
\alias{granger_test}
\title{Test de causalite de Granger}
\usage{
granger_test(object, cause, effect)
}
\arguments{
\item{object}{objet \code{var_fit}}

\item{cause}{indice de la variable causante}

\item{effect}{indice de la variable expliquee.}
}
\value{
liste : \code{statistic} (F), \code{df1}, \code{df2}, \code{p_value}.
}
\description{
\eqn{x} \strong{cause au sens de Granger} \eqn{y} si ses retards ameliorent la
prediction de \eqn{y}. Test F comparant l'equation de \eqn{y} avec et sans les
retards de \eqn{x}.
}
