% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_linalg.R
\name{householder_vector}
\alias{householder_vector}
\title{Vecteur de Householder (choix de signe stable)}
\usage{
householder_vector(x)
}
\arguments{
\item{x}{vecteur non nul.}
}
\value{
le vecteur v (même longueur que x) ; le vecteur nul si x est nul.
}
\description{
Renvoie le vecteur \eqn{v} définissant la réflexion \eqn{H(v)=I-2vv^T/v^Tv}
telle que \eqn{H(v)x = \alpha e_1}. Implémente l'éq. (0.5) : le signe
\eqn{v_1 = x_1 + \mathrm{sign}(x_1)\|x\|} évite l'annulation catastrophique.
}
