% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_cart.R
\name{impurity_entropy}
\alias{impurity_entropy}
\title{Entropie d'un vecteur d'étiquettes (éq. 8.1)}
\usage{
impurity_entropy(y)
}
\arguments{
\item{y}{étiquettes. @return \eqn{-\sum_c p_c\log p_c}.}
}
\description{
Entropie d'un vecteur d'étiquettes (éq. 8.1)
}
