% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_cart.R
\name{impurity_gini}
\alias{impurity_gini}
\title{Impureté de Gini d'un vecteur d'étiquettes (éq. 8.1)}
\usage{
impurity_gini(y)
}
\arguments{
\item{y}{étiquettes (facteur/vecteur). @return \eqn{1-\sum_c p_c^2}.}
}
\description{
Impureté de Gini d'un vecteur d'étiquettes (éq. 8.1)
}
