% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_cart.R
\name{impurity_variance}
\alias{impurity_variance}
\title{Impureté de variance (régression, éq. 8.2)}
\usage{
impurity_variance(y)
}
\arguments{
\item{y}{réponse numérique. @return variance de population intra-nœud.}
}
\description{
Impureté de variance (régression, éq. 8.2)
}
