% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_slt.R
\name{is_separable}
\alias{is_separable}
\title{Séparabilité linéaire d'un étiquetage (théorème de Gordan)}
\usage{
is_separable(X, y, tol = 1e-06)
}
\arguments{
\item{X}{matrice n x d des points.}

\item{y}{étiquettes dans \{-1, +1\} (longueur n).}

\item{tol}{seuil de distance pour déclarer 0 hors de l'enveloppe.}
}
\value{
TRUE si \{(x_i, y_i)\} est linéairement séparable (avec biais).
}
\description{
Séparabilité linéaire d'un étiquetage (théorème de Gordan)
}
