% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_nonparametric.R
\name{kde}
\alias{kde}
\title{Estimation de densite par noyau (KDE)}
\usage{
kde(x, grid, bw)
}
\arguments{
\item{x}{donnees}

\item{grid}{points d'evaluation}

\item{bw}{fenetre h}
}
\value{
vecteur des densites estimees sur \code{grid}.
}
\description{
\eqn{\hat f(x)=\frac1{nh}\sum_i K\!\bigl(\frac{x-x_i}{h}\bigr)}, noyau gaussien.
}
