% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_knn.R
\name{knn_regression}
\alias{knn_regression}
\title{Régression KNN (éq. 7.1)}
\usage{
knn_regression(X_train, y_train, X_test, k)
}
\arguments{
\item{X_train}{matrice n x p d'apprentissage.}

\item{y_train}{réponses d'apprentissage (numériques).}

\item{X_test}{matrice m x p de test.}

\item{k}{nombre de voisins.}
}
\value{
vecteur des m prédictions.
}
\description{
Prédit en chaque point de test la moyenne des réponses de ses k plus proches
voisins d'apprentissage.
}
