% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_generative.R
\name{lda_predict}
\alias{lda_predict}
\title{Prediction LDA}
\usage{
lda_predict(object, Xnew)
}
\arguments{
\item{object}{objet \code{lda_fit}}

\item{Xnew}{nouvelles observations.}
}
\value{
classes predites.
}
\description{
Prediction LDA
}
