% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_time_series.R
\name{ljung_box}
\alias{ljung_box}
\title{Test portmanteau de Ljung-Box (autocorrelation residuelle)}
\usage{
ljung_box(x, lag = 10L, fitdf = 0L)
}
\arguments{
\item{x}{serie (ou residus) ; @param lag nombre de retards K.}

\item{fitdf}{degres de liberte du modele ajuste (a soustraire).}
}
\value{
liste : \code{statistic}, \code{df}, \code{p_value}.
}
\description{
\eqn{Q=n(n+2)\sum_{k=1}^{K}\hat\rho_k^2/(n-k)\sim\chi^2_K} sous absence
d'autocorrelation. Utilise sur les residus d'un modele ajuste.
}
