% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/37_rnn.R
\name{lstm_forward}
\alias{lstm_forward}
\title{Passe avant d'une cellule LSTM (sequence)}
\usage{
lstm_forward(X, Wi, Wf, Wo, Wg, bi, bf, bo, bg)
}
\arguments{
\item{X}{sequence (T x d).}

\item{Wi, Wf, Wo, Wg}{matrices (H x (d+H)) empilant \eqn{[x_t,h_{t-1}]}.}

\item{bi, bf, bo, bg}{biais (H).}
}
\value{
liste : \code{H} (T x H), \code{C} (T x H).
}
\description{
Portes d'entree/oubli/sortie et etat de cellule :
\eqn{i=\sigma(W_i[x_t,h_{t-1}])}, \eqn{f=\sigma(\cdot)}, \eqn{o=\sigma(\cdot)},
\eqn{g=\tanh(\cdot)} ; \eqn{c_t=f\odot c_{t-1}+i\odot g}, \eqn{h_t=o\odot\tanh(c_t)}.
L'etat de cellule \eqn{c_t} forme une "autoroute" ou le gradient circule sans
s'evanouir.
}
