% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_m_estimation.R
\name{m_estimation_vcov}
\alias{m_estimation_vcov}
\title{Variance sandwich d'un M-estimateur (éq. 14.2)}
\usage{
m_estimation_vcov(psi, A)
}
\arguments{
\item{psi}{matrice n x p des valeurs de la fonction d'estimation \eqn{\psi_i}.}

\item{A}{matrice p x p, \eqn{\hat A} (p.ex. \eqn{X^\top X/n} pour l'OLS).}
}
\value{
la matrice de variance sandwich p x p de \eqn{\hat\theta}.
}
\description{
\eqn{\hat V = \hat A^{-1}\hat B\hat A^{-1}/n}, avec \eqn{\hat B = \frac1n\sum_i
\psi_i\psi_i^\top} et \eqn{\hat A} fournie (moyenne des dérivées de la fonction
d'estimation, \eqn{A = \mathbb E[-\partial_\theta\psi]}).
}
