% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28_autodiff.R
\name{Math.adnode}
\alias{Math.adnode}
\title{Fonctions élémentaires enregistrées (exp, log, sqrt, sin, cos, tanh)}
\usage{
\method{Math}{adnode}(x, ...)
}
\arguments{
\item{x}{\code{adnode}.}
}
\value{
\code{adnode}.
}
\description{
Fonctions élémentaires enregistrées (exp, log, sqrt, sin, cos, tanh)
}
