% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/36_cnn.R
\name{max_pool2d_backward}
\alias{max_pool2d_backward}
\title{Retropropagation du max-pooling}
\usage{
max_pool2d_backward(dout, cache)
}
\arguments{
\item{dout}{gradient en sortie (oh x ow) ; @param cache sortie de \code{max_pool2d}.}
}
\value{
\code{dX} (matrice de la taille de l'entree).
}
\description{
Route le gradient vers la position du maximum de chaque bloc (les autres
recoivent 0).
}
