% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_mlp.R
\name{mlp_backward}
\alias{mlp_backward}
\title{Rétropropagation : gradients analytiques (éq. 12.2-12.5)}
\usage{
mlp_backward(params, X, Y, activation, loss)
}
\arguments{
\item{params}{liste des paramètres.}

\item{X}{matrice n x d0.}

\item{Y}{cible n x d2.}

\item{activation}{activation cachée.}

\item{loss}{"mse" ou "logloss".}
}
\value{
liste des gradients \verb{W1, b1, W2, b2}.
}
\description{
Rétropropagation : gradients analytiques (éq. 12.2-12.5)
}
