% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_mlp.R
\name{mlp_forward}
\alias{mlp_forward}
\title{Passe avant du MLP (éq. 12.1)}
\usage{
mlp_forward(params, X, activation)
}
\arguments{
\item{params}{liste \verb{W1, b1, W2, b2}.}

\item{X}{matrice n x d0.}

\item{activation}{activation de la couche cachée ("tanh", "relu", "sigmoid").}
}
\value{
liste \verb{Z1, A1, Z2}.
}
\description{
Passe avant du MLP (éq. 12.1)
}
