% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_mlp.R
\name{mlp_loss}
\alias{mlp_loss}
\title{Perte moyenne du MLP}
\usage{
mlp_loss(Yhat, Y, loss)
}
\arguments{
\item{Yhat}{prédictions (sortie du réseau).}

\item{Y}{cible.}

\item{loss}{"mse" (régression) ou "logloss" (classification binaire).}
}
\value{
la perte moyenne.
}
\description{
Perte moyenne du MLP
}
