% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28_autodiff.R
\name{mm}
\alias{mm}
\title{Produit matriciel enregistré}
\usage{
mm(a, b)
}
\arguments{
\item{a, b}{\code{adnode} (ou matrices).}
}
\value{
\code{adnode}.
}
\description{
\eqn{C=AB} : \eqn{\bar A \mathrel{+}= \bar C B^\top}, \eqn{\bar B \mathrel{+}=
A^\top\bar C}.
}
