% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_cart.R
\name{n_leaves}
\alias{n_leaves}
\title{Nombre de feuilles d'un arbre CART}
\usage{
n_leaves(object)
}
\arguments{
\item{object}{objet \code{cart}. @return nombre de feuilles.}
}
\description{
Nombre de feuilles d'un arbre CART
}
