% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_nonparametric.R
\name{nadaraya_watson}
\alias{nadaraya_watson}
\title{Regression de Nadaraya-Watson (local constant)}
\usage{
nadaraya_watson(x, y, x0, bw)
}
\arguments{
\item{x, y}{donnees}

\item{x0}{points d'evaluation}

\item{bw}{fenetre}
}
\value{
vecteur des valeurs ajustees en \code{x0}.
}
\description{
\eqn{\hat m(x_0)=\sum_i K_h(x_0-x_i)\,y_i/\sum_i K_h(x_0-x_i)} : moyenne locale
ponderee par le noyau. Biais d'ordre \eqn{h^2}, mais biais de bord marque.
}
