% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_generative.R
\name{naive_bayes_predict}
\alias{naive_bayes_predict}
\title{Prediction Naive Bayes}
\usage{
naive_bayes_predict(object, Xnew)
}
\arguments{
\item{object}{objet \code{naive_bayes_fit} ; @param Xnew nouvelles observations.}
}
\value{
classes predites.
}
\description{
Prediction Naive Bayes
}
