% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ols.R
\name{ols_confint}
\alias{ols_confint}
\title{Intervalles de confiance des coefficients (éq. 1.8)}
\usage{
ols_confint(object, level = 0.95)
}
\arguments{
\item{object}{objet \code{ols}.}

\item{level}{niveau de confiance (défaut 0.95).}
}
\value{
matrice à deux colonnes (bornes inf/sup).
}
\description{
Intervalles de confiance des coefficients (éq. 1.8)
}
