% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ols.R
\name{ols_ftest}
\alias{ols_ftest}
\title{Test F de restrictions linéaires R beta = r (éq. 1.9)}
\usage{
ols_ftest(object, R, r = rep(0, nrow(R)))
}
\arguments{
\item{object}{objet \code{ols}.}

\item{R}{matrice q x p des restrictions (rang q).}

\item{r}{vecteur cible (longueur q ; défaut zéro).}
}
\value{
liste : \code{F}, \code{q}, \code{df.residual}, \code{p_value}.
}
\description{
Test F de restrictions linéaires R beta = r (éq. 1.9)
}
