% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_m_estimation.R
\name{ols_psi}
\alias{ols_psi}
\title{Fonction d'estimation de l'OLS : psi_i = x_i * e_i}
\usage{
ols_psi(X, resid)
}
\arguments{
\item{X}{matrice de design n x p (constante incluse).}

\item{resid}{résidus \eqn{\hat\varepsilon_i}.}
}
\value{
matrice n x p des \eqn{\psi_i = x_i \hat\varepsilon_i}.
}
\description{
Fonction d'estimation de l'OLS : psi_i = x_i * e_i
}
