% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28_autodiff.R
\name{Ops.adnode}
\alias{Ops.adnode}
\title{Opérateurs élémentaires enregistrés (+, -, *, /, ^)}
\usage{
\method{Ops}{adnode}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{adnode} ou numériques.}
}
\value{
\code{adnode}.
}
\description{
Opérateurs élémentaires enregistrés (+, -, *, /, ^)
}
