% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_time_series.R
\name{pacf_ts}
\alias{pacf_ts}
\title{Fonction d'autocorrelation partielle (PACF, recursion de Durbin-Levinson)}
\usage{
pacf_ts(x, lag.max = 10L)
}
\arguments{
\item{x}{serie ; @param lag.max retard maximal.}
}
\value{
vecteur des autocorrelations partielles (retards 1 a \code{lag.max}).
}
\description{
Fonction d'autocorrelation partielle (PACF, recursion de Durbin-Levinson)
}
