% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_quantile.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Perte pinball (fonction « check », éq. 20.1)}
\usage{
pinball_loss(u, tau)
}
\arguments{
\item{u}{résidu(s).}

\item{tau}{niveau de quantile dans (0,1).}
}
\value{
la (les) perte(s) pinball.
}
\description{
\eqn{\rho_\tau(u)=u(\tau-\mathbf 1\{u<0\})}. Vectorisée.
}
