% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_boosting.R
\name{predict_boost}
\alias{predict_boost}
\title{Prédiction d'un modèle de boosting}
\usage{
predict_boost(
  object,
  newdata,
  type = c("response", "link", "class"),
  n_trees = object$M
)
}
\arguments{
\item{object}{objet \code{boost}.}

\item{newdata}{data.frame des prédicteurs.}

\item{type}{"response" (valeur / probabilité), "link" (score F) ou "class"
(0/1, log-loss uniquement).}

\item{n_trees}{nombre d'arbres à utiliser (défaut : tous).}
}
\value{
vecteur de prédictions.
}
\description{
Prédiction d'un modèle de boosting
}
