% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_cart.R
\name{predict_cart}
\alias{predict_cart}
\title{Prédiction d'un arbre CART}
\usage{
predict_cart(object, newdata)
}
\arguments{
\item{object}{objet \code{cart}.}

\item{newdata}{data.frame contenant les prédicteurs.}
}
\value{
vecteur de prédictions (classes pour "class", moyennes pour "anova").
}
\description{
Prédiction d'un arbre CART
}
