% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_causal_ml.R
\name{predict_causal_tree}
\alias{predict_causal_tree}
\title{Prédiction du CATE par un arbre causal}
\usage{
predict_causal_tree(object, newdata)
}
\arguments{
\item{object}{objet \code{causal_tree}.}

\item{newdata}{covariables où prédire le CATE.}
}
\value{
vecteur des CATE prédits.
}
\description{
Prédiction du CATE par un arbre causal
}
