% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_bagging_rf.R
\name{predict_forest}
\alias{predict_forest}
\title{Prédiction d'une forêt / d'un ensemble baggé}
\usage{
predict_forest(object, newdata)
}
\arguments{
\item{object}{objet \code{forest}.}

\item{newdata}{data.frame des prédicteurs.}
}
\value{
vecteur de prédictions.
}
\description{
Agrège les B arbres : moyenne (régression) ou vote majoritaire (classification).
}
