% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_mlp.R
\name{predict_mlp}
\alias{predict_mlp}
\title{Prédiction d'un MLP}
\usage{
predict_mlp(model, newdata, type = c("response", "class"))
}
\arguments{
\item{model}{objet \code{mlp}.}

\item{newdata}{matrice des prédicteurs.}

\item{type}{"response" (valeur/probabilité) ou "class" (0/1, log-loss).}
}
\value{
vecteur/matrice de prédictions.
}
\description{
Prédiction d'un MLP
}
