% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_limited_dependent.R
\name{probit}
\alias{probit}
\title{Probit (reponse binaire) par IRLS (lien probit)}
\usage{
probit(formula, data, maxit = 50L)
}
\arguments{
\item{formula}{formule ; @param data data.frame ; @param maxit iterations max.}
}
\value{
liste : \code{coefficients}, \code{vcov}, \code{se}, \code{fitted}, \code{loglik}.
}
\description{
\eqn{P(y=1)=\Phi(X\beta)}. Maximum de vraisemblance par moindres carres
ponderes iteres (IRLS, Module 3) avec lien probit : poids
\eqn{w=\phi^2/[\Phi(1-\Phi)]}, reponse de travail \eqn{z=\eta+(y-\Phi)/\phi}.
}
