% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_generative.R
\name{qda_fit}
\alias{qda_fit}
\title{Analyse discriminante quadratique (QDA)}
\usage{
qda_fit(X, y)
}
\arguments{
\item{X, y}{cf. \code{lda_fit}.}
}
\value{
objet \code{qda_scratch}.
}
\description{
Chaque classe a sa \strong{propre} covariance \eqn{\Sigma_k} : frontiere
\strong{quadratique}. \eqn{\delta_k(x)=-\tfrac12\log|\Sigma_k|-\tfrac12(x-\mu_k)^\top
\Sigma_k^{-1}(x-\mu_k)+\log\pi_k}.
}
