% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_generative.R
\name{qda_predict}
\alias{qda_predict}
\title{Prediction QDA}
\usage{
qda_predict(object, Xnew)
}
\arguments{
\item{object}{objet \code{qda_fit}}

\item{Xnew}{nouvelles observations.}
}
\value{
classes predites.
}
\description{
Prediction QDA
}
