% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_diagnostics.R
\name{reset_test}
\alias{reset_test}
\title{Test RESET de Ramsey (forme fonctionnelle)}
\usage{
reset_test(formula, data, powers = 2:3)
}
\arguments{
\item{formula, data}{cf. \code{bp_test} ; @param powers puissances de \eqn{\hat y}.}
}
\value{
liste : \code{statistic} (F), \code{df1}, \code{df2}, \code{p_value}.
}
\description{
Ajoute des puissances des valeurs ajustees \eqn{\hat y^2,\hat y^3,\dots} au
modele et teste (F) leur significativite jointe. Rejet = non-linearite
negligee.
}
