% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/37_rnn.R
\name{rnn_backward}
\alias{rnn_backward}
\title{Retropropagation dans le temps (BPTT) du RNN simple}
\usage{
rnn_backward(dY, cache)
}
\arguments{
\item{dY}{gradient en sortie (T x O).}

\item{cache}{sortie de \code{rnn_forward}.}
}
\value{
liste : \code{dWxh}, \code{dWhh}, \code{dWhy}, \code{dbh}, \code{dby}.
}
\description{
Retropropagation dans le temps (BPTT) du RNN simple
}
