% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/37_rnn.R
\name{rnn_forward}
\alias{rnn_forward}
\title{Passe avant d'un RNN simple (sequence -> sequence)}
\usage{
rnn_forward(X, Wxh, Whh, Why, bh, by, h0 = rep(0, nrow(Wxh)))
}
\arguments{
\item{X}{sequence d'entree (T x d).}

\item{Wxh, Whh, Why}{matrices de poids (H x d), (H x H), (O x H).}

\item{bh, by}{biais (H), (O).}

\item{h0}{etat initial (defaut 0).}
}
\value{
liste : \code{Y} (T x O), \code{H} (T x H), \code{cache}.
}
\description{
\eqn{h_t=\tanh(W_{xh}x_t+W_{hh}h_{t-1}+b_h)}, \eqn{y_t=W_{hy}h_t+b_y}.
}
