% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_diagnostics.R
\name{sargan_test}
\alias{sargan_test}
\title{Test de suridentification de Sargan}
\usage{
sargan_test(y, X, Z)
}
\arguments{
\item{y}{reponse}

\item{X}{design (endogenes inclus)}

\item{Z}{instruments}
}
\value{
liste : \code{statistic}, \code{df}, \code{p_value}.
}
\description{
\eqn{n R^2} de la regression des residus 2SLS sur TOUS les instruments ;
\eqn{\sim\chi^2_{q-k}} (q instruments, k regresseurs). Rejet = instruments
invalides (correles au terme d'erreur).
}
