% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/38_attention.R
\name{softmax_rows}
\alias{softmax_rows}
\title{Softmax numeriquement stable (par ligne)}
\usage{
softmax_rows(X)
}
\arguments{
\item{X}{matrice ; @param axis 1 (par ligne, defaut).}
}
\value{
matrice de memes dimensions, chaque ligne sommant a 1.
}
\description{
Softmax numeriquement stable (par ligne)
}
