% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41_clustering.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Clustering spectral}
\usage{
spectral_clustering(X, k, gamma = 1)
}
\arguments{
\item{X}{matrice n x p}

\item{k}{nombre de clusters}

\item{gamma}{echelle RBF.}
}
\value{
vecteur d'assignation.
}
\description{
Construit un graphe de similarite (noyau RBF), calcule le \strong{Laplacien
normalise} \eqn{L=I-D^{-1/2}WD^{-1/2}}, plonge les points dans les \code{k} premiers
vecteurs propres, puis applique k-means. Separe des clusters \strong{non convexes}
(spirales, cercles) que le k-means echoue a distinguer.
}
