% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28_autodiff.R
\name{Summary.adnode}
\alias{Summary.adnode}
\title{Somme enregistrée (réduction scalaire)}
\usage{
\method{Summary}{adnode}(..., na.rm = FALSE)
}
\arguments{
\item{...}{un unique \code{adnode} à sommer.}

\item{na.rm}{ignoré (présent pour la signature du générique).}
}
\value{
\code{adnode} scalaire.
}
\description{
Somme enregistrée (réduction scalaire)
}
