% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/39_svm.R
\name{svm_rbf}
\alias{svm_rbf}
\title{Noyau gaussien (RBF) pour SVM}
\usage{
svm_rbf(X1, X2, gamma)
}
\arguments{
\item{X1, X2}{matrices}

\item{gamma}{largeur inverse.}
}
\value{
matrice de noyau.
}
\description{
Noyau gaussien (RBF) pour SVM
}
