% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_causal_ml.R
\name{t_learner}
\alias{t_learner}
\title{T-learner : CATE par deux modèles séparés}
\usage{
t_learner(X, y, d, newX = X, B = 200L)
}
\arguments{
\item{X}{data.frame de covariables (apprentissage).}

\item{y}{réponse.}

\item{d}{traitement (0/1).}

\item{newX}{covariables où prédire le CATE (défaut : X).}

\item{B}{nombre d'arbres.}
}
\value{
vecteur des CATE estimés en newX.
}
\description{
\eqn{\hat\tau(x) = \hat\mu_1(x) - \hat\mu_0(x)}, chaque \eqn{\mu} ajusté sur le
sous-groupe traité / contrôle (forêts du Module 9 par défaut).
}
