% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/44_var.R
\name{var_irf}
\alias{var_irf}
\title{Fonctions de reponse impulsionnelle (IRF) orthogonalisees}
\usage{
var_irf(object, h = 10L)
}
\arguments{
\item{object}{objet \code{var_fit} ; @param h horizon.}
}
\value{
tableau (h+1) x k x k : \code{irf[t, reponse, choc]}.
}
\description{
Reponse dynamique de chaque variable a un choc unitaire (orthogonalise par
Cholesky de \eqn{\Sigma}) dans chaque variable, sur \code{h} periodes. Coeur de
l'analyse macro-econometrique (propagation des chocs).
}
