% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_gls_robust.R
\name{vcov_nw}
\alias{vcov_nw}
\title{Variance HAC de Newey-West (éq. 2.6)}
\usage{
vcov_nw(fit, lag)
}
\arguments{
\item{fit}{objet \code{ols}.}

\item{lag}{nombre de retards L.}
}
\value{
matrice de variance p x p.
}
\description{
Viande \eqn{\hat\Gamma_0 + \sum_{\ell=1}^L w_\ell(\hat\Gamma_\ell +
\hat\Gamma_\ell^T)} avec poids de Bartlett \eqn{w_\ell = 1 - \ell/(L+1)}.
Calé sur \code{sandwich::NeweyWest(..., prewhite = FALSE, adjust = FALSE)}.
}
