% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_gls_robust.R
\name{wls_fit}
\alias{wls_fit}
\title{Moindres carrés pondérés (WLS, éq. 2.5)}
\usage{
wls_fit(formula, data, weights)
}
\arguments{
\item{formula}{formule façon \code{lm}.}

\item{data}{data.frame.}

\item{weights}{vecteur de poids \eqn{w_i = 1/\sigma_i^2} (longueur n).}
}
\value{
objet de classe \code{wls} (mêmes champs qu'\code{ols}).
}
\description{
Résout \eqn{(X^T W X)^{-1} X^T W y} en appliquant l'OLS (QR, Module 0) aux
données transformées \eqn{\sqrt{w_i}\,(x_i, y_i)}. Reproduit \code{lm(weights=)}.
}
