% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfromscratch-package.R
\docType{package}
\name{mlfromscratch-package}
\alias{mlfromscratch}
\alias{mlfromscratch-package}
\title{mlfromscratch : ML from Scratch in R}
\description{
Réimplémentation pédagogique, en R base, des principaux modèles d'apprentissage
automatique, adossée à leurs dérivations mathématiques et validée contre les
packages de référence. Voir les dérivations dans \verb{derivations/} et le rapport
de synthèse dans \verb{rapport/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mgourguechon.github.io/ml-from-scratch-R}
  \item \url{https://github.com/mgourguechon/ml-from-scratch-R}
  \item Report bugs at \url{https://github.com/mgourguechon/ml-from-scratch-R/issues}
}

}
\author{
\strong{Maintainer}: Maxime Gourguechon \email{maxime.gourguechon76@gmail.com}

Authors:
\itemize{
  \item Maxime Gourguechon \email{maxime.gourguechon76@gmail.com}
}

}
\keyword{internal}
